requirements.txt
# Production dependencies for TwitchClipper
selenium
moviepy
natsort
Pillow
pywin32
fastapi
uvicorn
webdriver-manager
